% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_ratio.R
\name{citation_ratio}
\alias{citation_ratio}
\title{Non profit & academic friendly ratio of citations}
\usage{
citation_ratio(doi)
}
\arguments{
\item{doi}{a \code{character} vector of Digital Object Identifiers (DOI). Can
contain \code{NA} (book, book chapter, etc.).}
}
\value{
A \code{list} of two elements:
\itemize{
\item \code{summary}, a \code{data.frame} with two columns (\code{metric} and \code{value})
reporting the following statistics:
\itemize{
\item number of total references (length of \code{doi} argument)
\item number of references with DOI
\item number of deduplicated references
\item number of references found in the OpenAlex database
\item number of references whose journal is indexed in the DAFNEE database
\item number of non-profit and academic friendly references
\item number of for-profit and academic friendly references
\item number of for-profit and non academic friendly references
}
\item \code{ratios}, a vector of three ratios:
\itemize{
\item non-profit and academic friendly ratio
\item for-profit and academic friendly ratio
\item for-profit and non academic friendly ratio
}
}
}
\description{
Scientific journals operate over a broad spectrum of publishing strategies,
from strictly for-profit, to non-profit, and in-between business models
(e.g. for-profit but academic friendly journals).

From a list of references, this function computes three citation ratios:
the proportion of non-profit citations, the proportion of for-profit and
academic friendly citations, and the proportion of for-profit and
non-academic friendly citations (Beck et al. 2025).

It uses the OpenAlex bibliographic database (\url{https://openalex.org}) to
retrieve journal names from article DOI and the DAFNEE database
(\url{https://dafnee.isem-evolution.fr/}) to get the business model and the
academic friendly status of journals.
}
\examples{
# Path to the BibTeX provided by <faircitations> ----
filename <- system.file(
  file.path("extdata", "references.bib"),
  package = "faircitations"
)

# Extract DOI from BibTeX ----
doi_list <- doi_from_bibtex(filename)

# Print DOI ----
doi_list

# Compute citation ratios ----
citation_ratio(doi_list)
}
\references{
Beck M et al. (2025) Strategic citations for a fairer academic
landscape. Submitted to Proc B - Biological Science Practices.
}
