% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_from_bibtex.R
\name{doi_from_bibtex}
\alias{doi_from_bibtex}
\title{Extract DOI from a BibTeX file}
\usage{
doi_from_bibtex(bibtex)
}
\arguments{
\item{bibtex}{a \code{character} of length 1. The (absolute or relative) path to
the BibTeX file to open.}
}
\value{
A \code{character} vector with extracted DOI. Some values can be \code{NA} in case of
books, chapters, etc. or if references are malformed.
}
\description{
This function reads a BibTex file and extracts the DOI of references (if
originally present in the file).
}
\examples{
# Path to the BibTeX provided by <faircitations> ----
filename <- system.file(
  file.path("extdata", "references.bib"),
  package = "faircitations"
)

# Extract DOI from BibTeX ----
doi_from_bibtex(filename)
}
