% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faircitations-package.R
\docType{package}
\name{faircitations-package}
\alias{faircitations}
\alias{faircitations-package}
\title{faircitations: How Fair Are You When You Cite Scientific Works?}
\description{
Provides a user-friendly way to compute the non-profit and academic friendly ratio of the bibliographic reference list before submitting a manuscript for peer review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/faircitations}
  \item Report bugs at \url{https://github.com/frbcesab/faircitations/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
