% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_get_dafnee_journals.R
\name{fp_get_dafnee_journals}
\alias{fp_get_dafnee_journals}
\title{List DAFNEE journals}
\usage{
fp_get_dafnee_journals()
}
\value{
A \code{data.frame} with three columns:
\itemize{
\item \code{journal}, the name of the journal
\item \code{business_model}, the business model of the journal (non-profit or
for-profit)
\item \code{academic_friendly}, the academic friendly status of the journal (yes or
no)
}
}
\description{
The DAFNEE database (Database of Academia‑Friendly Journals in Ecology and
Evolution, \url{https://dafnee.isem-evolution.fr/}) provides the business model
and the academic friendly status of several journals in the field of Ecology
and Evolution.

The \code{fairpub} package provides a selection of 291 DAFNEE journals and
this function returns information about these journals
}
\examples{
# List DAFNEE journals in fairpub ----
journals <- fp_get_dafnee_journals()

# Number of journals ----
nrow(journals)

# Preview of the outputs ----
head(journals)
}
