% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_ratio.R
\name{np_ratio}
\alias{np_ratio}
\title{Compute the non-profit ratio on a list of references}
\usage{
np_ratio(doi)
}
\arguments{
\item{doi}{a \code{character} vector of Digital Object Identifiers (DOI)}
}
\value{
A \code{data.frame} with two columns: \code{metric} and \code{value} reporting
different statistics.
}
\description{
...
}
\examples{
## ADD AN EXAMPLE
}
