% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faircitations-package.R
\docType{package}
\name{faircitations-package}
\alias{faircitations}
\alias{faircitations-package}
\title{faircitations: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/faircitations}
  \item Report bugs at \url{https://github.com/frbcesab/faircitations/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
