% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_ratio.R
\name{np_ratio}
\alias{np_ratio}
\title{Compute the non-profit ratio on a list of references}
\usage{
np_ratio(doi)
}
\arguments{
\item{doi}{a \code{character} vector of Digital Object Identifiers (DOI)}
}
\value{
A \code{data.frame} with two columns: \code{metric} and \code{value} reporting
different statistics.
}
\description{
...
}
\examples{
# Import a BiBTex file (provided example) ----
filename <- system.file(
  file.path("extdata", "references.bib"),
  package = "faircitations"
)

refs <- RefManageR::ReadBib(filename)

length(refs)
refs[1:5]

# Extract DOI ----
doi_list <- unlist(refs$"doi")
names(doi_list) <- NULL
doi_list

# Compute NP ratio ----
np_ratio(doi_list)
}
